+++ packages/core/src/languages/html/common.ts
@@ -1,11 +1,12 @@
 /**
- * Get a Regular Expression pattern as a string for certain attributes in HTML.
+ * Get a Regular Expression pattern as a string for certain attributes with
+ * quoted values in HTML.
  *
  * @param attributesPattern The pattern of the attributes to match.
  * @param quotePattern The pattern of the quote(s) to match.
  * @returns A Regular Expression pattern as a string.
  */
-export const ATTRIBUTE_PATTERN = (
+export const QUOTED_ATTRIBUTE_PATTERN = (
   attributesPattern: string,
   quotePattern: string,
 ): string => `\\s(?:${attributesPattern})\\s*=\\s*${quotePattern}\\s*`;
@@ -13,7 +14,7 @@ export const ATTRIBUTE_PATTERN = (
 /**
  * Array of characters that are quotes in HTML.
  */
-export const QUOTES_ARRAY = ["\"", "'"];
+export const QUOTES_ARRAY: string[] = ["\"", "'"];
 
 /**
  * Regular Expression pattern as a string for quotes in HTML.