+++ packages/core/src/languages/html/single-value-attributes.ts
@@ -2,31 +2,30 @@ import type { MangleExpression } from "../../types";
 import type { SingleValueAttributeOptions } from "../options";
 
 import { SingleGroupMangleExpression } from "../utils/mangle-expressions";
-import { ATTRIBUTE_PATTERN, QUOTES_PATTERN } from "./common";
+import { QUOTED_ATTRIBUTE_PATTERN, QUOTES_PATTERN } from "./common";
 
 const GROUP_MAIN = "main";
 const GROUP_QUOTE = "quote";
 
 /**
- * Get a {@link MangleExpression} to match element attribute values in HTML,
- * e.g. `bar` in `<div data-foo="bar"></div>` or `sun` in
- * `<div data-praise="thesun"></div>` if the value prefix is "the".
+ * Get a {@link MangleExpression} to match quoted element attribute values in
+ * HTML, e.g. `bar` in `<div data-foo="bar"></div>` or `sun` in `<div data-
+ * praise="thesun"></div>` if the value prefix is "the".
  *
- * @param attributeNames A list of attribute names.
- * @param valuePrefix An expression of the required prefix for values.
- * @param valueSuffix An expression of the required suffix for values.
- * @returns The {@link MangleExpression} to match attribute values in HTML.
+ * @param attributesPattern The pattern of attribute names.
+ * @param valuePrefix An pattern of the required prefix for values.
+ * @param valueSuffix An pattern of the required suffix for values.
+ * @returns The {@link MangleExpression} to match quoted attribute values.
  */
-function newElementAttributeSingleValueExpression(
-  attributeNames: string[],
+function newQuotedValueExpression(
+  attributesPattern: string,
   valuePrefix: string,
   valueSuffix: string,
 ): MangleExpression {
-  const attributesPattern = attributeNames.join("|");
   return new SingleGroupMangleExpression(
     `
       (?<=
-        ${ATTRIBUTE_PATTERN(
+        ${QUOTED_ATTRIBUTE_PATTERN(
           attributesPattern,
           `(?<${GROUP_QUOTE}>${QUOTES_PATTERN})`,
         )}
@@ -43,25 +42,56 @@ function newElementAttributeSingleValueExpression(
 }
 
 /**
+ * Get a {@link MangleExpression} to match unquoted element attribute values in
+ * HTML, e.g. `bar` in `<div data-foo=bar></div>` or `sun` in `<div data-
+ * praise=thesun></div>` if the value prefix is "the".
+ *
+ * @param attributesPattern The pattern of attribute names.
+ * @param valuePrefix An pattern of the required prefix for values.
+ * @param valueSuffix An pattern of the required suffix for values.
+ * @returns The {@link MangleExpression} to match unquoted attribute values.
+ */
+function newUnquotedValueExpression(
+  attributesPattern: string,
+  valuePrefix: string,
+  valueSuffix: string,
+): MangleExpression {
+  return new SingleGroupMangleExpression(
+    `
+      (?<=
+        \\s(?:${attributesPattern})\\s*=
+        ${valuePrefix}
+      )
+      (?<${GROUP_MAIN}>%s)
+      (?=
+        ${valueSuffix}
+        (?:\\s|\\/|\\>)
+      )
+    `,
+    GROUP_MAIN,
+  );
+}
+
+/**
  * Get the set of {@link MangleExpression}s to match single-value attribute
  * values in HTML. This will match:
- * - Attribute values (e.g. `bar` in `<div data-foo="bar"></div>`).
+ * - Quoted attribute values (e.g. `bar` in `<div data-foo="bar"></div>`).
+ * - Unquoted attribute values (e.g. `bar` in `<div data-foo=bar></div>`).
  *
  * @param options The {@link SingleValueAttributeOptions}.
  * @returns A set of {@link MangleExpression}s.
  * @since v0.1.14
+ * @version v0.1.16
  */
 export default function singleValueAttributeExpressionFactory(
   options: SingleValueAttributeOptions,
 ): MangleExpression[] {
+  const attributesPattern = options.attributeNames.join("|");
   const valuePrefix = options.valuePrefix ? options.valuePrefix : "";
   const valueSuffix = options.valueSuffix ? options.valueSuffix : "";
 
   return [
-    newElementAttributeSingleValueExpression(
-      options.attributeNames,
-      valuePrefix,
-      valueSuffix,
-    ),
+    newQuotedValueExpression(attributesPattern, valuePrefix, valueSuffix),
+    newUnquotedValueExpression(attributesPattern, valuePrefix, valueSuffix),
   ];
 }