+++ packages/core/src/languages/javascript/common.ts
@@ -1,7 +1,7 @@
 /**
  * Array of characters that are quotes in JavaScript.
  */
-export const QUOTES_ARRAY = ["\"", "'", "`"];
+export const QUOTES_ARRAY: string[] = ["\"", "'", "`"];
 
 /**
  * Regular Expression pattern as a string for quotes in JavaScript.